Finishing up initial YouTube scraping
Added helper functions for extracting text and saving the current casper
page in helper.js. Linked these in main and test. urlmp3 writes an array
of song objects into an html file with proper anchor tags. test
currently contains code for extracting title, artist and itunes link (if
it exists). If itunes exists, this is working perfectly. Need to clean
the data for the other case.