Works for url and bad_url
Insted of having a common thenOpen after the if-else, I have a thenOpen
only in the if. In the else I have put a this.click() which basically
clicks the first link in the google-search. For some reason, after
getting the iTunes url in the else, the original common thenOpen was NOT
opening the corresponding iTunes page. Also, this.click() seems to be
cleaner and more natural